feat: render book card
