+++ src/pages/HomePage/HomePage.tsx
@@ -1,11 +1,7 @@
 import React, { Component } from 'react';
 import SearchPanel from './components/SearchPanel/SearchPanel';
 import BookCardList from './components/BookCardList/BookCardList';
-import {
-  BookCard,
-  BookCardResponseBooks,
-  BookCardResponseWorks,
-} from './components/BookCard/models';
+import { BookCard, BooksResponse } from './components/BookCard/models';
 
 interface State {
   searchTerm: string;
@@ -17,13 +13,7 @@ export default class App extends Component<object, State> {
     super(props);
     this.state = {
       searchTerm: '',
-      searchResults: [
-        {
-          id: '1',
-          title: '"Fantastic Mr Fox"',
-          cover: '6498519',
-        },
-      ],
+      searchResults: [],
     };
   }
 
@@ -33,7 +23,14 @@ export default class App extends Component<object, State> {
     )
       .then((res) => res.json())
       .then((data: BooksResponse) =>
-        this.setState({ searchResults: data.docs as any })
+        this.setState({
+          searchResults: data.docs.map((doc) => ({
+            id: doc.key,
+            title: doc.title,
+            cover: doc.cover_i?.toString() || '',
+            authors: doc.author_name || [],
+          })),
+        })
       )
       .catch((err) => console.log(err));
   };
@@ -50,22 +47,4 @@ export default class App extends Component<object, State> {
       </div>
     );
   }
-}
-// Generated by https://quicktype.io
-
-export interface BooksResponse {
-  numFound: number;
-  start: number;
-  numFoundExact: boolean;
-  docs: Doc[];
-  num_found: number;
-  q: string;
-  offset: null;
-}
-
-export interface Doc {
-  key: string;
-  title: string;
-  cover_i?: number;
-  author_name: string[];
 }
