+++ src/pages/HomePage/components/BookCardList/BookCardList.tsx
@@ -24,6 +24,7 @@ class BookCardList extends Component<BookCardListProps> {
               <div className="card-body">
                 <h5 className="card-title">{book.title}</h5>
                 <h5 className="card-cover">{book.cover}</h5>
+                <h5 className="card-author">{book.authors}</h5>
               </div>
             </div>
           );
